refactor(indicator): move classes to models.py
move metadata and result class to indicators/models.py
